unified testing crates

+++ test-schemas/src/tests/schema_tests/rendering_tests.rs
@@ -119,7 +119,10 @@ fn rendering_test() {
   let pkg = RENDERING_PKG::get_package();
 
   let output1 = PathBuf::from(concat!(env!("CARGO_MANIFEST_DIR"), "/rendering_test"));
-  fs::remove_dir_all(&output1).unwrap();
+
+  if output1.exists() {
+    fs::remove_dir_all(&output1).unwrap();
+  }
 
   pkg.render_files(&output1).unwrap();
 
@@ -145,7 +148,10 @@ fn rendering_test() {
   let manual_pkg = Package::new("rendering").with_files([ManualFile::file_schema()]);
 
   let output2 = PathBuf::from(concat!(env!("CARGO_MANIFEST_DIR"), "/rendering_test2"));
-  fs::remove_dir_all(&output2).unwrap();
+
+  if output2.exists() {
+    fs::remove_dir_all(&output2).unwrap();
+  }
 
   manual_pkg.render_files(&output2).unwrap();
